# swagger/swagger.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UsersController_create","summary":"Создать пользователя","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}}},"responses":{"204":{"description":""}},"tags":["Управление пользователями"]},"get":{"operationId":"UsersController_getAll","summary":"Получить список пользователей","parameters":[{"name":"limit","required":true,"in":"query","example":200000,"description":"Лимит","schema":{"minimum":1,"default":200000,"type":"number"}},{"name":"offset","required":true,"in":"query","example":0,"description":"Смещение","schema":{"minimum":0,"default":0,"type":"number"}},{"name":"sort","required":false,"in":"query","description":"Столбец сортировки","schema":{"default":"login","type":"string"}},{"name":"order","required":false,"in":"query","description":"Сортировка. Направление. \"ASC\" - прямое, \"DESC\" - обратное.","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"search","required":false,"in":"query","description":"Поисковая строка. Поиск по имени, фамилии, email","example":"mail.ru","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserList"}}}}},"tags":["Управление пользователями"]}},"/users/{login}":{"put":{"operationId":"UsersController_update","summary":"Изменить пользователя","parameters":[{"name":"login","required":true,"in":"path","example":"admin","description":"логин","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"204":{"description":""}},"tags":["Управление пользователями"]},"delete":{"operationId":"UsersController_delete","summary":"Удалить пользователя","parameters":[{"name":"login","required":true,"in":"path","example":"admin","description":"логин","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Управление пользователями"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"Залогиниться (access токен устанавливается в сервер куку!!!)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginParams"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JWT_Refresh"}}}}},"tags":["Авторизация"]}},"/auth/logout/{login}":{"post":{"operationId":"AuthController_logout","summary":"Разлогиниться","parameters":[{"name":"login","required":true,"in":"path","example":"admin","description":"логин","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Авторизация"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","summary":"Получить новый access токен","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JWT_Refresh"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JWT_Refresh"}}}}},"tags":["Авторизация"]}},"/auth":{"get":{"operationId":"AuthController_whoAmi","summary":"Получить информацию о пользователе по access токену","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserItem"}}}}},"tags":["Авторизация"]}}},"info":{"title":"Authority service openApi","description":"API Docs","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserCreate":{"type":"object","properties":{"login":{"type":"string","example":"admin","description":"логин"},"firstName":{"type":"string","example":"Иван","description":"имя"},"lastName":{"type":"string","example":"Иванов","description":"фамилия"},"email":{"type":"string","example":"example@gmail.com","description":"Почта"},"role":{"type":"string","example":"admin","description":"роль","enum":["admin","operator","guest"]},"locked":{"type":"boolean","example":true,"description":"Заблокирован - true/ разблокирован - false"},"password":{"type":"string","example":"384633ad37a18b3b4fc5bf3e371d6e9f","description":"Пароль+логин в md5"}},"required":["login","firstName","lastName","email","role","locked"]},"UserUpdate":{"type":"object","properties":{"firstName":{"type":"string","example":"Иван","description":"имя"},"lastName":{"type":"string","example":"Иванов","description":"фамилия"},"email":{"type":"string","example":"example@gmail.com","description":"Почта"},"role":{"type":"string","example":"admin","description":"роль","enum":["admin","operator","guest"]},"locked":{"type":"boolean","example":true,"description":"Заблокирован - true/ разблокирован - false"},"password":{"type":"string","example":"384633ad37a18b3b4fc5bf3e371d6e9f","description":"Пароль+логин в md5"}}},"UserItem":{"type":"object","properties":{"login":{"type":"string","example":"admin","description":"логин"},"firstName":{"type":"string","example":"Иван","description":"имя"},"lastName":{"type":"string","example":"Иванов","description":"фамилия"},"email":{"type":"string","example":"example@gmail.com","description":"Почта"},"role":{"type":"string","example":"admin","description":"роль","enum":["admin","operator","guest"]},"locked":{"type":"boolean","example":true,"description":"Заблокирован - true/ разблокирован - false"},"createdAt":{"format":"date-time","type":"string","description":"Время создания"},"updatedAt":{"format":"date-time","type":"string","description":"Время обновления"}},"required":["login","firstName","lastName","email","role","locked"]},"UserList":{"type":"object","properties":{"items":{"example":[],"description":"Список пользователей","type":"array","items":{"$ref":"#/components/schemas/UserItem"}},"count":{"type":"number","example":0,"description":"Всего пользователей соответсвуют фильтру"}},"required":["items","count"]},"LoginParams":{"type":"object","properties":{"login":{"type":"string","example":"admin","description":"логин"},"sessionId":{"type":"string","example":"-","default":"-","description":"ID сессии"},"password":{"type":"string","example":"3ZRwlSi7HIPQjzCI1AQ/R0KJH08=","description":"Пароль+логин в md5 (в примере логин - admin пароль - 12345678 md5sum(admin12345678))"}},"required":["login"]},"JWT_Refresh":{"type":"object","properties":{"refreshToken":{"type":"string","description":"refresh токен"}},"required":["refreshToken"]}}}}